feat: Expand integer literal parsing to i128 in the lexer and add parser validation for i64 and usize bounds.

+++ src/lexer/tokens.rs
@@ -5,7 +5,7 @@ use crate::utils::Span;
 #[derive(Debug, Clone, PartialEq)]
 pub enum TokenKind {
     // Literals
-    Int(i64),
+    Int(i128),
     Float(f64),
     Str(String),
     FString(String),
